probe caches: store grids bf16 (half host-RAM/IO) so M2 fits 1 GPU
The row-subset fix (b734ed7) did NOT get M2 onto a 1-GPU slot: dense-label
tasks (onset/speech) keep ~all rows, so grid[keep] still copies the full
55 GB fp32 electrode grid + the mmap pages it -> ~110 GB > the 111 GB slot
(measured OOM at 112/116 GB). bf16 storage halves both -> ~55 GB -> 1 GPU.

- save_cache casts float grids to bf16 (store_dtype, default bf16). Re-cast
  an existing cache = save_cache(load_cache(p), p).
- AttentiveProbeHead.forward upcasts x to fp32 -> compute (and AUROC)
  unchanged; only per-minibatch is upcast, storage stays bf16.
- feature_matrix upcasts to fp32 for the numpy ridge path.
- tests: roundtrip asserts bf16 dtype + AUROC within bf16 tol; new
  attentive-on-bf16 recovers signal.

+++ src/speech_decoding/experiments/online_probe.py
@@ -104,7 +104,7 @@ def feature_matrix(pooled: Tensor, parcels: Tensor | tp.Sequence[int]) -> Tensor
     """Select ``parcels`` from ``(B, n_parcels, F, k, d)`` and flatten → ``(B, P·F·k·d)``."""
     idx = torch.as_tensor(list(parcels), dtype=torch.long, device=pooled.device)
     sel = pooled.index_select(1, idx)
-    return sel.reshape(sel.shape[0], -1)
+    return sel.reshape(sel.shape[0], -1).float()   # bf16 caches -> fp32 for the numpy ridge
 
 
 def parcel_intersection(present_a: Tensor, present_b: Tensor) -> Tensor:
